trading_bot/requirements.txt
ccxt==4.2.25
pandas>=2.2.0
numpy>=1.26.4
python-dotenv==1.0.0
websockets==12.0
supabase==2.3.4
ta==0.11.0
scikit-learn==1.4.0
scipy==1.12.0
plotly==5.18.0
requests==2.31.0
aiohttp==3.9.1
